+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.12.1
+
+#### FIXES
+- Within Update `comment` on `dependencytrack_team_apikey` was not being written back to state using return value from DependencyTrack.
+- Witin Create, Update, the permission list was assigned from a `dtrack.Team`, which may have been `nil`.
+
+#### MISC
+- Standardised log error reporting when failing to parse a `UUID`.
+
 ## 1.12.0
 
 #### FEATURES
@@ -62,6 +71,9 @@
 
 #### FEATURES
 - Add `dependencytrack_team_permissions` resource to canonically manage the permissions assigned to a Team.
+
+#### ISSUES
+- [Fixed in `1.12.1`]: Permission list is assigned from a potentially `nil` `dtrack.Team`.
 
 #### MISC
 - Remove deprecated field from within `golangci` config file for `goconst`.